pole angle calculation
Fixed pole angle calculation to use spherical trig.  Importantly, the
haversine formula is used instead of the law of cosines, because it is
not affected by small angles.  This assumes that the azimuthal pointing
angle was originally calculated using spherical trig (which is the case
for SuperDARN).  This assumption makes it possible to make no
assumptions as to the nature of the vector quantity being calculated.